% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{combine_nest}
\alias{combine_nest}
\title{Combine columns and nest data for each permutation}
\usage{
combine_nest(.data, .names_from, .values_from)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.names_from}{The columns to build the permutations on (e.g., c(col1, col2))}

\item{.values_from}{The columns to be nested for each permutation (e.g., c(col3, col4, col5))}
}
\value{
A nested `tbl`
}
\description{
combine_nest() takes as imput a `tbl` formatted as | <element> | <feature> | <value> | <...> | and returns a `tbl` with data nested for each permutation.
}
\details{
\lifecycle{maturing}

...
}
\examples{

combine_nest(mtcars_tidy, car_model, c(feature,value))


}
