% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_histogram}
\alias{create_histogram}
\title{Creates an histogram from a data frame}
\usage{
create_histogram(.data, .abundance)
}
\arguments{
\item{.data}{A numerical vector}

\item{.abundance}{A column name for the variable to be plotted}
}
\value{
A `ggplot` object
}
\description{
create_histogram() creates a `ggplot` object from a data frame
}
\details{
\lifecycle{experimental}

This function creates a ggpot histogram from a data frame
}
\examples{

create_histogram(
   tibble::tibble(counts = 1:100),
   counts
)

}
