% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fill_missing,tbl_df-method}
\alias{fill_missing,tbl_df-method}
\title{fill_missing}
\usage{
\S4method{fill_missing}{tbl_df}(.data, .element = NULL, .feature = NULL, .value = NULL, fill_with)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.element}{The name of the element column}

\item{.feature}{The name of the feature/gene column}

\item{.value}{The name of the feature/gene value column}

\item{fill_with}{A numerical value with which fill the mssing data points}
}
\value{
A `tbl` with filld abundnce
}
\description{
fill_missing
}
