% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_NA_using_formula}
\alias{fill_NA_using_formula}
\title{This function is needed for DE in case the matrix is not rectangular, but includes NA}
\usage{
fill_NA_using_formula(
  .data,
  .formula,
  .element = NULL,
  .feature = NULL,
  .value = NULL,
  .value_scaled = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{a formula with no response variable, of the kind ~ factor_of_intrest + batch}

\item{.element}{The name of the element column}

\item{.feature}{The name of the feature/gene column}

\item{.value}{The name of the feature/gene value column}

\item{.value_scaled}{The name of the feature/gene scaled value column}
}
\value{
A tibble with adjusted counts
}
\description{
This function is needed for DE in case the matrix is not rectangular, but includes NA
}
