% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{gate_dimensions}
\alias{gate_dimensions}
\alias{gate_dimensions,spec_tbl_df-method}
\alias{gate_dimensions,tbl_df-method}
\title{Label points within a scatter plot drawing a gate}
\usage{
gate_dimensions(
  .data,
  .element,
  .dim1,
  .dim2,
  name = "inside_gate",
  action = "add",
  ...
)

\S4method{gate_dimensions}{spec_tbl_df}(
  .data,
  .element,
  .dim1,
  .dim2,
  name = "inside_gate",
  action = "add",
  ...
)

\S4method{gate_dimensions}{tbl_df}(
  .data,
  .element,
  .dim1,
  .dim2,
  name = "inside_gate",
  action = "add",
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{name}{A character string. The name of the new column}

\item{action}{A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tbl object with additional columns for the inside gate information. additional columns for the rotated dimensions. The rotated dimensions will be added to the original data set as `<NAME OF DIMENSION> rotated <ANGLE>` by default, or as specified in the input arguments.

A tbl object with additional columns for the reduced dimensions. additional columns for the rotated dimensions. The rotated dimensions will be added to the original data set as `<NAME OF DIMENSION> rotated <ANGLE>` by default, or as specified in the input arguments.

A tbl object with additional columns for the reduced dimensions. additional columns for the rotated dimensions. The rotated dimensions will be added to the original data set as `<NAME OF DIMENSION> rotated <ANGLE>` by default, or as specified in the input arguments.
}
\description{
gate_dimensions() takes as input a `tbl` formatted as | <DIMENSION 1> | <DIMENSION 2> | <...> | and calculates the rotated dimensional space of the feature value.
}
\details{
\lifecycle{maturing}

This function allow the user to label data points in inside a 2D gate.
}
\examples{

\donttest{

 mtcars_tidy_MDS = reduce_dimensions(mtcars_tidy, car_model, feature, value, method="MDS")
 
 gate_dimensions(mtcars_tidy_MDS, car_model, `Dim1`, `Dim2`)
 
}

}
