% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gate_dimensions_}
\alias{gate_dimensions_}
\title{Get points within a user drawn gate}
\usage{
gate_dimensions_(
  .data,
  .element,
  .dim1,
  .dim2,
  .color = NULL,
  .shape = NULL,
  .size = NULL,
  name = "inside_gate",
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tibble with additional columns
}
\description{
Get points within a user drawn gate
}
\keyword{internal}
