% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gate_list}
\alias{gate_list}
\title{Gate list for clustering}
\format{
An object of class \code{list} of length 2.
}
\usage{
gate_list
}
\description{
Gate list for clustering
}
\keyword{internal}
