% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_clusters_kmeans_bulk}
\alias{get_clusters_kmeans_bulk}
\title{Get K-mean clusters to a tibble}
\usage{
get_clusters_kmeans_bulk(
  .data,
  .element = NULL,
  .feature = NULL,
  .value = NULL,
  of_elements = TRUE,
  transform = NULL,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally elements)}

\item{.value}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{of_elements}{A boolean}

\item{transform}{A function to use to tranforma the data internalli (e.g., log1p)}

\item{...}{Further parameters passed to the function kmeans}
}
\value{
A tibble with additional columns
}
\description{
Get K-mean clusters to a tibble
}
