% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_reduced_dimensions_TSNE_bulk}
\alias{get_reduced_dimensions_TSNE_bulk}
\title{Get principal component information to a tibble using tSNE}
\usage{
get_reduced_dimensions_TSNE_bulk(
  .data,
  .element = NULL,
  .feature = NULL,
  .value = NULL,
  .dims = 2,
  top = Inf,
  of_elements = TRUE,
  transform = NULL,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally elements)}

\item{.feature}{A column symbol. The column that is represents entities to cluster (i.e., normally genes)}

\item{.value}{A column symbol with the value the clustering is based on (e.g., `count`)}

\item{.dims}{A integer vector corresponding to principal components of interest (e.g., 1:6)}

\item{top}{An integer. How many top genes to select}

\item{of_elements}{A boolean}

\item{transform}{A function to use to tranforma the data internalli (e.g., log1p)}

\item{...}{Further parameters passed to the function Rtsne}
}
\value{
A tibble with additional columns
}
\description{
Get principal component information to a tibble using tSNE
}
