% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_x_y_annotation_columns}
\alias{get_x_y_annotation_columns}
\title{get_x_y_annotation_columns}
\usage{
get_x_y_annotation_columns(.data, .horizontal, .vertical, .value)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.horizontal}{The name of the column horizontally presented in the heatmap}

\item{.vertical}{The name of the column vertically presented in the heatmap}

\item{.value}{The name of the feature/gene value column}
}
\value{
A list
}
\description{
This function recognise what are the element-wise columns and transcrip-wise columns
}
