% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{impute_missing,tbl_df-method}
\alias{impute_missing,tbl_df-method}
\title{impute_missing}
\usage{
\S4method{impute_missing}{tbl_df}(
  .data,
  .formula,
  .element = NULL,
  .feature = NULL,
  .value = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.element}{The name of the element column}

\item{.feature}{The name of the feature/gene column}

\item{.value}{The name of the feature/gene value column}
}
\value{
A `tbl` with imputed abundnce
}
\description{
impute_missing
}
