% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{keep_variable,spec_tbl_df-method}
\alias{keep_variable,spec_tbl_df-method}
\title{keep_variable}
\usage{
\S4method{keep_variable}{spec_tbl_df}(.data, .names_from, .values_from)
}
\arguments{
\item{.data}{A `tbl`}
}
\value{
A `tbl` with filld abundnce
}
\description{
keep_variable
}
