% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keep_variable_features}
\alias{keep_variable_features}
\title{Identify variable genes for dimensionality reduction}
\usage{
keep_variable_features(
  .data,
  .element = NULL,
  .feature = NULL,
  .value = NULL,
  top = Inf,
  transform = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A character name of the element column}

\item{.feature}{A character name of the transcript/gene column}

\item{.value}{A character name of the read count column}

\item{top}{An integer. How many top genes to select}

\item{transform}{A function to use to tranforma the data internalli (e.g., log1p)}
}
\value{
A tibble filtered genes
}
\description{
Identify variable genes for dimensionality reduction
}
