% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{lower_triangular}
\alias{lower_triangular}
\title{Keep rows corresponding of a lower triangular matrix built from two columns}
\usage{
lower_triangular(.data, .col1, .col2, .value)
}
\arguments{
\item{.data}{A `tbl`}

\item{.col1}{A column name}

\item{.col2}{A column name}
}
\value{
A `tbl` with filtered rows
}
\description{
lower_triangular() takes as imput a `tbl` formatted as | <element> | <feature> | <value> | <...> | and returns a filtered `tbl`
}
\details{
\lifecycle{maturing}

...
}
\examples{


res =
	lower_triangular(
		nanny::counts_mini,
	~ condition,
	.element = element,
	.feature = feature,
	.value = count
)


}
