% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lower_triangular,spec_tbl_df-method}
\alias{lower_triangular,spec_tbl_df-method}
\title{lower_triangular}
\usage{
\S4method{lower_triangular}{spec_tbl_df}(.data, .col1, .col2, .value)
}
\arguments{
\item{.data}{A `tbl`}

\item{.col1}{A column name}

\item{.col2}{A column name}
}
\value{
A `tbl` with filld abundnce
}
\description{
lower_triangular
}
