% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_tidy}
\alias{mtcars_tidy}
\title{Cibersort reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 288 rows and 5 columns.
}
\usage{
mtcars_tidy
}
\description{
Cibersort reference
}
\keyword{datasets}
