% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nest_subset}
\alias{nest_subset}
\title{Nest according to selected-column-wise information}
\usage{
nest_subset(.data, ...)
}
\arguments{
\item{.data}{A `tbl`}

\item{.column}{The name of the column of interest}
}
\value{
A `tbl` object
}
\description{
nest_subset() takes as imput a `tbl` and returns a nested `tbl` according to only selected-column-related columns
}
\details{
\lifecycle{maturing}

This functon extracts only selected-column-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to nanny function.
}
\examples{

nest_subset(mtcars_tidy,data = -car_model)


}
