% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nest_subset,tbl-method}
\alias{nest_subset,tbl-method}
\title{nest_subset}
\usage{
\S4method{nest_subset}{tbl}(.data, ..., .names_sep = NULL)
}
\arguments{
\item{.data}{A `tbl`}

\item{...}{The name of the columns of interest}

\item{.names_sep}{Deprecated by tidyr}
}
\value{
A `tbl` object
}
\description{
nest_subset
}
