% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nest_subset,tbl_df-method}
\alias{nest_subset,tbl_df-method}
\title{nest_subset}
\usage{
\S4method{nest_subset}{tbl_df}(.data, ..., .names_sep = NULL)
}
\arguments{
\item{.data}{A `tbl`}
}
\value{
A `tbl` object
}
\description{
nest_subset
}
