% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nest_subset.tbl}
\alias{nest_subset.tbl}
\title{nest_subset}
\usage{
\method{nest_subset}{tbl}(.data, ..., .exclude = NULL, .names_sep = NULL)
}
\arguments{
\item{.data}{A `tbl`}

\item{...}{The name of the columns of interest}

\item{.exclude}{Column name. It is the column\(s\) that you might want to exclude from the subset.}

\item{.names_sep}{Deprecated by tidyr}
}
\value{
A `tbl` object
}
\description{
nest_subset
}
