% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{permute_nest}
\alias{permute_nest}
\title{Permute columns and nest data for each permutation}
\usage{
permute_nest(.data, .names_from, .values_from)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.names_from}{The columns to build the permutations on (e.g., c(col1, col2))}

\item{.values_from}{The columns to be nested for each permutation (e.g., c(col3, col4, col5))}
}
\value{
A nested `tbl`
}
\description{
permute_nest() takes as imput a `tbl` formatted as | <element> | <feature> | <value> | <...> | and returns a `tbl` with data nested for each permutation.
}
\details{
\lifecycle{maturing}

...
}
\examples{


res =
	permute_nest(
		nanny::counts_mini,
	~ condition,
	.element = element,
	.feature = feature,
	.value = count
)


}
