% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{permute_nest,tbl_df-method}
\alias{permute_nest,tbl_df-method}
\title{permute_nest}
\usage{
\S4method{permute_nest}{tbl_df}(.data, .names_from, .values_from)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.names_from}{The columns to build the permutations on (e.g., c(col1, col2))}

\item{.values_from}{The columns to be nested for each permutation (e.g., c(col3, col4, col5))}
}
\value{
A `tbl` with filld abundnce
}
\description{
permute_nest
}
