% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_redundancy,tbl_df-method}
\alias{remove_redundancy,tbl_df-method}
\title{remove_redundancy}
\usage{
\S4method{remove_redundancy}{tbl_df}(
  .data,
  .element = NULL,
  .feature = NULL,
  .value = NULL,
  of_elements = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  transform = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.element}{The name of the element column (normally elements).}

\item{.feature}{The name of the feature column (normally features)}

\item{.value}{The name of the column including the numerical value the clustering is based on (normally feature value)}

\item{of_elements}{A boolean. In case the input is a nanny object, it indicates Whether the element column will be element or feature column}

\item{correlation_threshold}{A real number between 0 and 1. For correlation based calculation.}

\item{top}{An integer. How many top genes to select for correlation based method}

\item{transform}{A function to use to tranforma the data internalli (e.g., log1p)}
}
\value{
A tbl object with with dropped recundant elements (e.g., elements).
}
\description{
remove_redundancy
}
