% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset,tbl_df-method}
\alias{subset,tbl_df-method}
\title{subset}
\usage{
\S4method{subset}{tbl_df}(.data, .column)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}
}
\value{
A `tbl` object
}
\description{
subset
}
