% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data2}
\alias{test_data2}
\title{Example data set reduced}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2240 rows and 6 columns.
}
\usage{
test_data2
}
\description{
Example data set reduced
}
\keyword{internal}
