% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_matrix}
\alias{as_matrix}
\title{Get matrix from tibble}
\usage{
as_matrix(.data, rownames = NULL, do_check = TRUE, sep_rownames = "___")
}
\arguments{
\item{rownames}{A character string of the rownames}

\item{do_check}{A boolean}

\item{tbl}{A tibble}
}
\value{
A matrix
}
\description{
Get matrix from tibble
}
\examples{

as_matrix(head(dplyr::select(nanny::counts_mini, feature, count)), rownames=feature)

}
