% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_matrix,spec_tbl_df-method}
\alias{as_matrix,spec_tbl_df-method}
\title{as_matrix}
\usage{
\S4method{as_matrix}{spec_tbl_df}(.data, rownames = NULL, do_check = TRUE, sep_rownames = "___")
}
\arguments{
\item{rownames}{A character string of the rownames}

\item{do_check}{A boolean}
}
\value{
A `tbl` with filld abundnce
}
\description{
as_matrix
}
