% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{impute_missing}
\alias{impute_missing}
\title{Impute feature value if missing from element-feature pairs}
\usage{
impute_missing(
  .data,
  .formula,
  .element = NULL,
  .feature = NULL,
  .value = NULL
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.formula}{A formula with no response variable, representing the desired linear model where the first covariate is the factor of interest and the second covariate is the unwanted variation (of the kind ~ factor_of_intrest + batch)}

\item{.element}{The name of the element column}

\item{.feature}{The name of the feature/gene column}

\item{.value}{The name of the feature/gene value column}
}
\value{
A `tbl` non-sparse value
}
\description{
impute_missing() takes as imput a `tbl` formatted as | <element> | <feature> | <value> | <...> | and returns a `tbl` with an edditional adjusted value column. This method uses scaled counts if present.
}
\details{
\lifecycle{maturing}

This function imputes the value of missing element-feature pair using the median of the element group defined by the formula
}
\examples{


res =
	impute_missing(
		nanny::counts_mini,
	~ condition,
	.element = element,
	.feature = feature,
	.value = count
)


}
