% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{lower_triangular}
\alias{lower_triangular}
\title{Keep rows corresponding of a lower triangular matrix built from two columns}
\usage{
lower_triangular(.data, .col1, .col2, .value)
}
\arguments{
\item{.data}{A `tbl`}

\item{.col1}{A column name}

\item{.col2}{A column name}

\item{.value}{A column names of the value column}
}
\value{
A `tbl` with filtered rows
}
\description{
lower_triangular() takes as imput a `tbl` formatted as | <element> | <feature> | <value> | <...> | and returns a filtered `tbl`
}
\details{
\lifecycle{maturing}

...
}
\examples{

library(dplyr)
library(purrr)
library(tidyr)

mtcars_tidy_permuted = permute_nest(mtcars_tidy, car_model, c(feature,value))

mtcars_tidy_permuted \%>\%
 # Summarise mpg
 mutate(data = map(data, ~ .x \%>\% filter(feature == "mpg") \%>\% summarise(mean(value)))) \%>\%
unnest(data) \%>\%

# Lower triangular
lower_triangular(car_model_1, car_model_2,  `mean(value)`)

}
