% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{subset}
\alias{subset}
\title{Extract sampe-wise information}
\usage{
subset(.data, .column)
}
\arguments{
\item{.data}{A `tbl` formatted as | <element> | <feature> | <value> | <...> |}

\item{.column}{The name of the element column}
}
\value{
A `tbl` object
}
\description{
subset() takes as imput a `tbl` formatted as | <element> | <ENSEMBL_ID> | <value> | <...> | and returns a `tbl` with only sampe-related columns
}
\details{
\lifecycle{maturing}

This functon extracts only element-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to nanny function.
}
\examples{

subset(mtcars_tidy,car_model)


}
